# Backend/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PMS Admin API",
    "description": "Auto-generated API documentation for PMS Admin Portal",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000/"
    }
  ],
  "paths": {
...
    "/api/plans/plans": {
      "post": {
        "tags": [
          "Plans"
        ],
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "product_id": {
                    "example": "any"
                  },
                  "plan_name": {
                    "example": "any"
                  },
                  "title": {
                    "example": "any"
                  },
                  "description": {
                    "example": "any"
                  },
                  "including_ids": {
                    "example": "any"
                  },
                  "excluding_ids": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/plansByProduct/{product_id}": {
      "get": {
        "tags": [
          "Plans"
        ],
        "description": "",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/plans/plans/{id}": {
      "put": {
        "tags": [
          "Plans"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "product_id": {
                    "example": "any"
                  },
                  "plan_name": {
                    "example": "any"
                  },
                  "title": {
                    "example": "any"
                  },
                  "description": {
                    "example": "any"
                  },
                  "including_ids": {
                    "example": "any"
                  },
                  "excluding_ids": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Plans"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/plans/planPoints": {
      "post": {
        "tags": [
          "Plans"
        ],
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "point_name": {
                    "example": "any"
                  },
                  "icon": {
                    "example": "any"
                  },
                  "description": {
                    "example": "any"
                  },
                  "type": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Plans"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/plans/assignPointsToPlan": {
      "post": {
        "tags": [
          "Plans"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "plan_id": {
                    "example": "any"
                  },
                  "point_ids": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/planPoints/{plan_id}": {
      "get": {
        "tags": [
          "Plans"
        ],
        "description": "",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/plans/promoCode": {
      "post": {
        "tags": [
          "Plans"
        ],
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "plan_id": {
                    "example": "any"
                  },
                  "promo_code": {
                    "example": "any"
                  },
                  "unit_type": {
                    "example": "any"
                  },
                  "unit_value": {
                    "example": "any"
                  },
                  "valid_from": {
                    "example": "any"
                  },
                  "valid_to": {
                    "example": "any"
                  },
                  "usage_limit": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/promoCode/{id}": {
      "put": {
        "tags": [
          "Plans"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "plan_id": {
                    "example": "any"
                  },
                  "promo_code": {
                    "example": "any"
                  },
                  "unit_type": {
                    "example": "any"
                  },
                  "unit_value": {
                    "example": "any"
                  },
                  "valid_from": {
                    "example": "any"
                  },
                  "valid_to": {
                    "example": "any"
                  },
                  "usage_limit": {
                    "example": "any"
                  },
                  "is_active": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/promoCodeToggle/{id}": {
      "put": {
        "tags": [
          "Plans"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/plans/planRate": {
      "post": {
        "tags": [
          "Plans"
        ],
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "plan_id": {
                    "example": "any"
                  },
                  "tenure_type": {
                    "example": "any"
                  },
                  "display_price": {
                    "example": "any"
                  },
                  "selling_price": {
                    "example": "any"
                  },
                  "is_discountable": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/planRate/{id}": {
      "put": {
        "tags": [
          "Plans"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "plan_id": {
                    "example": "any"
                  },
                  "tenure_type": {
                    "example": "any"
                  },
                  "display_price": {
                    "example": "any"
                  },
                  "selling_price": {
                    "example": "any"
                  },
                  "is_discountable": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/planRateToggle/{id}": {
      "put": {
        "tags": [
          "Plans"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}